requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
scikit-learn
pandas
